requirements.txt
APScheduler==3.6.3
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
decorator==4.4.2
google-auth==1.24.0
google-auth-oauthlib==0.4.2
gspread==3.6.0
httplib2==0.18.1
idna==2.10
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dateutil==2.8.1
python-decouple==3.4
python-telegram-bot==13.1
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
